+++ Backend/package.json
@@ -22,33 +22,30 @@
     "morgan": "^1.10.0",
     "nanoid": "^5.1.5",
     "nodemon": "^3.1.9",
-
-    "ts-node": "^10.9.2",
-    "typescript": "^5.8.3",
-
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3",
     "uuid": "^11.1.0",
-
     "zod": "^3.24.2"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
     "@types/cookie-parser": "^1.4.8",
     "@types/express": "^5.0.1",
+    "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^9.0.9",
     "@types/morgan": "^1.9.9",
     "@types/node": "^22.14.1",
+    "@types/supertest": "^6.0.3",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-express": "^4.1.8",
-    "rimraf": "^6.0.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.8.3",
-    "@types/supertest": "^6.0.3",
     "@types/uuid": "^10.0.0",
     "jest": "^29.7.0",
+    "rimraf": "^6.0.1",
     "supertest": "^7.1.0",
     "ts-jest": "^29.3.1",
-    "@types/jest": "^29.5.14"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
   }
 }
