+++ Backend/src/models/asset/assetModel.ts
@@ -1,6 +1,7 @@
 import { Document, model, Schema } from "mongoose";
 
-//I am inside the models and it looks great
+// I am inside the models and it looks great
+
 interface IAsset extends Document {
   assetId: string;
   name: string;
@@ -17,12 +18,12 @@ interface IAsset extends Document {
   createdBy: string;
   createdAt: Date;
   updatedAt: Date;
-  isDeleted: boolean;
+  isDeleted: Boolean;
 }
 
 const assetSchema = new Schema<IAsset>(
   {
-    assetId: { type: String, required: true },
+    assetId: { type: String, required: true, unique: true },
     name: { type: String, required: true },
     category: { type: String, required: true },
     description: { type: String },
@@ -35,17 +36,12 @@ const assetSchema = new Schema<IAsset>(
     assignedTo: { type: String },
     purchaseDate: { type: Date },
     warrantyDate: { type: Date },
-    lastMaintained: { type: Date },
     nextMaintenance: { type: Date },
     qrCodeUrl: { type: String },
     createdBy: { type: String, required: true },
-    // createdAt: { type: Date, default: Date.now },
-    // updatedAt: { type: Date, default: Date.now },
     isDeleted: { type: Boolean, default: false },
   },
-  {
-    timestamps: true,
-  }
+  { timestamps: true }
 );
 
 export const Asset = model<IAsset>("Asset", assetSchema);
