+++ Backend/src/validators/asset/asset.validator.ts
@@ -1,11 +1,13 @@
 import { z } from "zod";
 
+const AssetStatus = z.enum(["available", "in_use", "maintenance", "retired"]);
+
 export const assetSchema = z.object({
   name: z.string().min(1),
   category: z.string().min(1),
   description: z.string().optional(),
   location: z.string().min(1),
-  status: z.enum(["available", "in_use", "maintenance", "retired"]),
+  status: AssetStatus,
   assignedTo: z.string().optional(),
   purchaseDate: z.coerce.date().optional(),
   warrantyDate: z.coerce.date().optional(),
@@ -14,3 +16,5 @@ export const assetSchema = z.object({
   qrCodeUrl: z.string().url().optional(),
   createdBy: z.string().min(1),
 });
+
+export const updateSchema = assetSchema.partial();
